+++ NEWS.md
@@ -5,6 +5,10 @@
 * Export `dplyr::n()`, to meet forthcoming changes in dplyr 0.8.0.
 * `plot_model()` and `tab_model()` now support `MixMod`-objects from package **GLMMadpative**, `mlogit`- and `gmnl`-models.
 
+## Changes to functions
+
+* `tab_model()` gets a `strings`-argument, as convenient shortcut for setting column-header strings.
+
 ## Bug fixes
 
 * Legend labels were inverted for **brms**-models in `plot_model(..., type = "diag")`.